la till valdering för alla metoder, skapade validator metod för price

+++ src/user/user.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, InternalServerErrorException, NotFoundException } from '@nestjs/common';
+import { ForbiddenException, Injectable, InternalServerErrorException, NotFoundException } from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { EditUserDto } from './dto/edit.user.dto';
 
@@ -12,6 +12,17 @@ export class UserService {
     dto: EditUserDto,
   ) {
     try {
+      if (!Number.isInteger(userId) || userId <= 0) {
+        throw new ForbiddenException('ogitligt användare id, det måste vara ett positivt nummer');
+      }
+  
+      const existUser = await this.prisma.user.findUnique({
+        where: {id: userId},
+      });
+  
+      if (!existUser) {
+        throw new ForbiddenException(`ingen användare med angivna id: ${userId}`);
+      }
         const user = await this.prisma.user.update({
           where: {
             id: userId.id,
@@ -24,13 +35,26 @@ export class UserService {
         delete user.hash;
         return user;
     } catch (error){
-        console.error('Error updating user:', error);
-        throw new InternalServerErrorException('Failed to update user');
+        console.error('Fel vid uppdatering av användare:', error);
+        throw new InternalServerErrorException('Det gick inte att uppdatera användare');
     }
 
   }
 
   async deleteUser(userId: number) {
+
+    if (!Number.isInteger(userId) || userId <= 0) {
+      throw new ForbiddenException('ogitligt användare id, det måste vara ett positivt nummer');
+    }
+
+    const existUser = await this.prisma.user.findUnique({
+      where: {id: userId},
+    });
+
+    if (!existUser) {
+      throw new ForbiddenException(`ingen användare med angivna id: ${userId}`);
+    }
+
     const deleteUser = await this.prisma.user.delete({
       where: {
         id: userId,
